Make code more consistent
