+++ sql/catalyst/src/main/scala/org/apache/spark/sql/util/SchemaUtils.scala
@@ -17,7 +17,8 @@
 
 package org.apache.spark.sql.util
 
-import org.apache.spark.internal.Logging
+import org.apache.spark.sql.AnalysisException
+import org.apache.spark.sql.types.StructType
 
 
 /**
@@ -25,29 +26,26 @@ import org.apache.spark.internal.Logging
  *
  * TODO: Merge this file with [[org.apache.spark.ml.util.SchemaUtils]].
  */
-private[spark] object SchemaUtils extends Logging {
+private[spark] object SchemaUtils {
 
-  /**
-   * Checks if input column names have duplicate identifiers. Prints a warning message if
-   * the duplication exists.
-   *
-   * @param columnNames column names to check
-   * @param colType column type name, used in a warning message
-   * @param caseSensitiveAnalysis whether duplication checks should be case sensitive or not
-   */
+  def checkSchemaColumnNameDuplication(
+      schema: StructType, colType: String, caseSensitiveAnalysis: Boolean = false): Unit = {
+    checkColumnNameDuplication(schema.map(_.name), colType, caseSensitiveAnalysis)
+  }
+
   def checkColumnNameDuplication(
-      columnNames: Seq[String], colType: String, caseSensitiveAnalysis: Boolean): Unit = {
-    val names = if (caseSensitiveAnalysis) {
-      columnNames
+      names: Seq[String], colType: String, caseSensitiveAnalysis: Boolean = false): Unit = {
+    val colNames = if (caseSensitiveAnalysis) {
+      names
     } else {
-      columnNames.map(_.toLowerCase)
+      names.map(_.toLowerCase)
     }
-    if (names.distinct.length != names.length) {
-      val duplicateColumns = names.groupBy(identity).collect {
-        case (x, ys) if ys.length > 1 => s"`$x`"
+    if (colNames.distinct.length != colNames.length) {
+      val duplicateColumns = colNames.groupBy(identity).collect {
+        case (x, ys) if ys.length > 1 => "\"" + x + "\""
       }
-      logWarning(s"Found duplicate column(s) $colType: ${duplicateColumns.mkString(", ")}. " +
-        "You might need to assign different column names.")
+      throw new AnalysisException(s"Found duplicate column(s) in $colType: " +
+        duplicateColumns.mkString(", "))
     }
   }
 }
