+++ sql/catalyst/src/test/scala/org/apache/spark/sql/util/SchemaUtilsSuite.scala
