+++ sql/core/src/main/scala/org/apache/spark/sql/execution/command/ddl.scala
@@ -430,7 +430,8 @@ case class AlterTableAddPartitionCommand(
         spec,
         table.partitionColumnNames,
         table.identifier.quotedString,
-        sparkSession.sessionState.conf.resolver)
+        sparkSession.sessionState.conf.resolver,
+        sparkSession.sessionState.conf.caseSensitiveAnalysis)
       // inherit table storage format (possibly except for location)
       CatalogTablePartition(normalizedSpec, table.storage.copy(
         locationUri = location.map(CatalogUtils.stringToURI)))
@@ -480,13 +481,15 @@ case class AlterTableRenamePartitionCommand(
       oldPartition,
       table.partitionColumnNames,
       table.identifier.quotedString,
-      sparkSession.sessionState.conf.resolver)
+      sparkSession.sessionState.conf.resolver,
+      sparkSession.sessionState.conf.caseSensitiveAnalysis)
 
     val normalizedNewPartition = PartitioningUtils.normalizePartitionSpec(
       newPartition,
       table.partitionColumnNames,
       table.identifier.quotedString,
-      sparkSession.sessionState.conf.resolver)
+      sparkSession.sessionState.conf.resolver,
+      sparkSession.sessionState.conf.caseSensitiveAnalysis)
 
     catalog.renamePartitions(
       tableName, Seq(normalizedOldPartition), Seq(normalizedNewPartition))
@@ -528,7 +531,8 @@ case class AlterTableDropPartitionCommand(
         spec,
         table.partitionColumnNames,
         table.identifier.quotedString,
-        sparkSession.sessionState.conf.resolver)
+        sparkSession.sessionState.conf.resolver,
+        sparkSession.sessionState.conf.caseSensitiveAnalysis)
     }
 
     catalog.dropPartitions(
